Mark all protocol logic classes as internal and move to new Io namespace
